[Windows only] -remove code to turn off screensaver due to problems -updated version number to b33 -added icon for add/remove programs list [All platforms] -for the flock size in the F2 display, if playing gold sheep then report gold flock size -shorten delay when no sheep are available -show native string for open file in log

+++ client_generic/ContentDownloader/Shepherd.h
@@ -91,10 +91,10 @@ class Shepherd
 	static bool getSheep( char *path, SheepArray *sheep );
 
 	static uint64 s_ClientFlockBytes;
-	static size_t s_ClientFlockCount;
+	static uint64 s_ClientFlockCount;
 
 	static uint64 s_ClientFlockGoldBytes;
-	static size_t s_ClientFlockGoldCount;
+	static uint64 s_ClientFlockGoldCount;
 
 	static char *fRootPath;
 	static char *fMpegPath;
@@ -365,7 +365,7 @@ class Shepherd
 					return s_ClientFlockGoldBytes/1024/1024;
 				return 0;
 			}
-			static size_t getClientFlockCount(const int generationtype)
+			static uint64 getClientFlockCount(const int generationtype)
 			{
 				if ( generationtype == 0 )
 					return s_ClientFlockCount;
